+++ CHANGELOG.md
@@ -6,9 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ## [v0.6.0] - 2021-06-24
 ### Added
-- Threaded Async Executor for smol. ([#254](https://github.com/paritytech/substrate-archive/commit/36d955d379b1fdfb0ff063dce394d8a4d6430323))
+- Use `async-std` instead of `smol` for async tasks. ([#301](https://github.com/paritytech/substrate-archive/commit/f2a4b408123c5e64be04e70a890b1354475d812d))
 - Additional tests for missing storage & pagination query. ([#254](https://github.com/paritytech/substrate-archive/commit/36d955d379b1fdfb0ff063dce394d8a4d6430323))
 - Storage Chain Support ([#292](https://github.com/paritytech/substrate-archive/commit/70ea92187ed68aee9c9cc8de757b1b86bb7ab408))
 - `v0.9.5`, `v0.9.6` wasm tracing runtimes
@@ -23,6 +24,11 @@ where lots of blocks were missing from storage at one time. ([#254](https://gith
 
 ### Removed
 - **BREAKING** `db_actor_pool_size` is no longer configurable on the builder. Actor pool no longer exists. ([#254](https://github.com/paritytech/substrate-archive/commit/36d955d379b1fdfb0ff063dce394d8a4d6430323))
+- **BREAKING** `max_tasks` from builder. Max tasks will always be the number of CPUS on the system. ([#301](https://github.com/paritytech/substrate-archive/commit/f2a4b408123c5e64be04e70a890b1354475d812d))
+- **BREAKING** `task_workers` from builder. Task workers will always be number of CPUS on the system. ([#301](https://github.com/paritytech/substrate-archive/commit/f2a4b408123c5e64be04e70a890b1354475d812d))
+
+### Fixed
+- archive no longer hangs on Ctrl-C. ([#301](https://github.com/paritytech/substrate-archive/commit/f2a4b408123c5e64be04e70a890b1354475d812d))
 
 ## [v0.5.2] - 2021-06-02
 ### Added